+++ src/containers/App.js
@@ -1,6 +1,9 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import {Link, IndexLink} from 'react-router';
+
+import Header from './Header/Header';
+import SideMenu from './SideMenu/SideMenu';
+import MainWrapper from './MainWrapper/MainWrapper';
 
 // This is a class-based component because the current
 // version of hot reloading won't hot reload a stateless
@@ -8,15 +11,11 @@ import {Link, IndexLink} from 'react-router';
 class App extends React.Component {
 	render() {
 		return (
-			<div>
-				<IndexLink to="/">Home</IndexLink>
-				{' | '}
-				<Link to="/fuel-savings">Demo App</Link>
-				{' | '}
-				<Link to="/about">About</Link>
-				<br/>
-				{this.props.children}
-			</div>
+			<main>
+				<Header />
+				<SideMenu />
+				<MainWrapper>{this.props.children}</MainWrapper>
+			</main>
 		);
 	}
 }