+++ CHANGELOG.md
@@ -7,6 +7,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - Project setup
 - Setting the global style (.editorconfig) to use tabs with 4 spaces and reformatting code
 - Added versioning.txt and CHANGELOG.md to track changes and versions appropriately
+- Added container components - Header, SideMenu and MainWrapper
 
 
 ## TODO